move header to own component, update styles
